A demo
The APP should try not to restore from a process kill so that it will
not have a chance to use any lost data kept in the static variables
(the content in the heap of previously killed process).

Instead, it should start over.

Then a program should be able to use static variables to keep its state
without crashes.